+++ pydantic_xml/__init__.py
@@ -6,4 +6,4 @@ from . import config, errors
 from .errors import ModelError
 from .model import BaseGenericXmlModel, BaseXmlModel, XmlAttributeInfo, XmlElementInfo, XmlWrapperInfo, attr, element
 from .model import wrapped
-from .serializers import DEFAULT_ENCODER, XmlEncoder
+from .serializers.encoder import DEFAULT_ENCODER, XmlEncoder
