+++ pydantic_xml/serializers/__init__.py
@@ -1,0 +1,2 @@
+from . import factories
+from .encoder import DEFAULT_ENCODER, XmlEncoder
